Refactor: move PigeonMaps.kt JS module to separate package
Since this is not part of material-ui components
